Fixing compiler warnings.

+++ Lumberjack/DDTTYLogger.m
@@ -90,7 +90,7 @@ static DDTTYLogger *sharedInstance;
 	if (logMsg)
 	{
 		const char *msg = [logMsg UTF8String];
-		int msgLen = strlen(msg);
+		size_t msgLen = strlen(msg);
 		
 		if (isFormatted)
 		{
@@ -113,7 +113,7 @@ static DDTTYLogger *sharedInstance;
 			NSString *tsNStr = [dateFormatter stringFromDate:(logMessage->timestamp)];
 			
 			const char *tsCStr = [tsNStr UTF8String];
-			int tsLen = strlen(tsCStr);
+			size_t tsLen = strlen(tsCStr);
 			
 			// tid = thread id
 			// 
@@ -149,7 +149,7 @@ static DDTTYLogger *sharedInstance;
 			v[5].iov_len = 1;
 			
 			v[6].iov_base = tidCStr;
-			v[6].iov_len = MIN(8, tidLen); // snprintf doesn't return what you might think
+			v[6].iov_len = MIN((size_t)8, tidLen); // snprintf doesn't return what you might think
 			
 			v[7].iov_base = "] ";
 			v[7].iov_len = 2;
